+++ src/datarobot_genai/drmcp/core/middleware.py
@@ -26,16 +26,16 @@ from starlette.requests import Request
 from starlette.responses import JSONResponse
 from starlette.responses import Response
 
-from datarobot_genai.core.exceptions import AudienceClaimValidationError
-from datarobot_genai.core.utils.auth import JWTTokenClaimsValidator
 from datarobot_genai.drmcp.core.config import get_config
 from datarobot_genai.drmcpbase.middleware import AuthContextExtractor
 from datarobot_genai.drmcpbase.middleware import OAuthMiddleWare
 from datarobot_genai.drmcpbase.middleware import register_oauth_middleware
+from datarobot_genai.drmcputils.auth import JWTTokenClaimsValidator
 from datarobot_genai.drmcputils.auth import extract_auth_context_from_headers
 from datarobot_genai.drmcputils.auth import set_auth_context
 from datarobot_genai.drmcputils.auth import set_request_headers
 from datarobot_genai.drmcputils.constants import AUTH_CTX_KEY
+from datarobot_genai.drmcputils.exceptions import AudienceClaimValidationError
 
 from .routes_utils import prefix_mount_path